requirements.txt
appnope==0.1.2
backcall==0.2.0
decorator==5.1.0
ipython==7.30.1
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
Pygments==2.11.0
traitlets==5.1.1
wcwidth==0.2.5
